messed thing up refactored wrong component in wrong folder fixed now
