+++ src/components/index.js
@@ -11,11 +11,6 @@ export { default as Container } from './Container';
  * (aka original)
  --------------------------- */
 
-/** Still needs to be refactored into functional */
-/** HAVING ISSUES WITH THE REFACTORED VERSIONS OF: */
-// export { default as MTableEditField } from './m-table-edit-field';
-export { default as MTableEditCell } from './m-table-edit-cell';
-
 /** ---------------------------
  * Functional components
  * (aka refactor)
@@ -23,6 +18,7 @@ export { default as MTableEditCell } from './m-table-edit-cell';
 export { default as MTableBodyRow } from './MTableBodyRow';
 export { default as MTableBody } from './MTableBody';
 export { default as MTableEditField } from './MTableEditField';
+export { default as MTableEditCell } from './MTableEditCell';
 
 // Trying to keep these in alphabetical order
 export { default as MTableAction } from './MTableAction';
